requirements.txt
Django==1.6.2
Mezzanine==3.1.9
MySQL-python==1.2.3
PAM==0.4.2
Pillow==2.3.0
South==1.0
Twisted-Core==13.2.0
Twisted-Web==13.2.0
Unidecode==0.04.16
adium-theme-ubuntu==0.3.4
appdirs==1.3.0
apt-xapian-index==0.45
argparse==1.2.1
beautifulsoup4==4.1.3
bleach==1.4
chardet==2.0.1
colorama==0.2.5
command-not-found==0.3
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.10
django-ajax==2.2.10
django-appconf==0.6
django-autoslug==1.7.2
django-cities==0.4.1
django-cities-light==2.4.2
django-classy-tags==0.4
django-countries==2.1.2
django-geoposition==0.2.2
django-google-maps==0.2.3
django-localflavor-in==1.0
django-location-field==1.3.2
django-pagination==1.0.7
django-social-auth==0.7.28
django-twitter-tag==1.2
duplicity==0.6.23
filebrowser-safe==0.3.5
future==0.9.0
futures==2.1.6
gmusicapi==4.0.0
grappelli-safe==0.3.12
html5lib==0.999
httplib2==0.8
lockfile==0.8
lxml==3.3.3
mock==1.0.1
mutagen==1.23
numpy==1.8.1
oauth2==1.5.211
oauth2client==1.2
oauthlib==0.6.3
oneconf==0.3.7
pexpect==3.1
piston-mini-client==0.7.5
proboscis==1.2.6.0
protobuf==2.5.0
pyOpenSSL==0.13
pycrypto==2.6.1
pycups==1.9.66
pycurl==7.19.3
pygobject==3.12.0
pyinotify==0.9.4
pyserial==2.6
pysmbc==1.0.14.1
python-apt==0.9.3.5
python-dateutil==2.2
python-debian==0.1.21-nmu2ubuntu2
python-openid==2.2.5
pytz==2014.4
pyxdg==0.25
reportlab==3.0
requests==2.3.0
requests-oauthlib==0.4.1
rhythmbox-gmusic==4.0dev
sessioninstaller==0.0.0
six==1.2.0
software-center-aptd-plugins==0.0.0
system-service==0.1.6
twitter==1.9.1
tzlocal==1.0
unity-lens-photos==1.0
urllib3==1.7.1
validictory==0.9.3
wsgiref==0.1.2
xdiagnose==3.6.3build2
zope.interface==4.0.5
